[CH2] refactor : use JUnit
