+++ BE/Spring/토비의 스프링 3.1/Vol.1/springbook/build.gradle
@@ -13,11 +13,12 @@ dependencies {
 
     implementation 'org.springframework:spring-context:5.3.22'
     implementation 'org.springframework:spring-jdbc:5.3.22'
+    testImplementation 'org.springframework:spring-test:5.3.22'
 
     implementation('org.projectlombok:lombok:1.18.26')
     implementation('mysql:mysql-connector-java:8.0.33')
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
+    implementation platform('org.junit:junit-bom:5.9.1')
+    implementation 'org.junit.jupiter:junit-jupiter'
 }
 
 test {
