+++ BE/Spring/토비의 스프링 3.1/Vol.1/springbook/src/main/java/org/example/user/dao/UserDAO.java
@@ -14,7 +14,6 @@ public class UserDAO {
     private User user;
 
     public UserDAO(DataSource dataSource) {
-
         this.dataSource = dataSource;
     }
